'notice'

+++ tensquare_parent/tensquare_common/src/main/java/com/tensquare/entity/PageResult.java
@@ -4,15 +4,15 @@ import java.util.List;
 
 public class PageResult<T> {
 
-    private int total;
+    private long total;
     private List<T> rows;
 
-    public PageResult(int total, List<T> rows) {
+    public PageResult(long total, List<T> rows) {
         this.total = total;
         this.rows = rows;
     }
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
